experiment/requirements.txt
numpy
pandas
Pillow
scikit-learn == 1.2.0
streamlit == 1.20.0